Add script to render sprites to PNG
When updating the sprites, we need to run this again in order for Godot to have the correctly rendered sprites.
I'm hesitant to put the output of this script in the repository. But I guess I'll have to eventually if I'm ever intending to share this with other computers (inc. my own laptop) easily.