Add spawner
This is very simplistic. Normally the mobs would get added to the scene rather than to this spawner. The spawn counts and timouts and composition would have to be loaded from a configuration file. The spawn points would have to be spread out over multiple spawn locations and perhaps randomised for flying units. All that later.